feat: Add timeout for blpop

+++ src/main/java/redis/RedisHandler.java
@@ -44,7 +44,7 @@ public class RedisHandler {
         validateNumberOfArgs(req, 2);
         var key = req.get(1);
         var timeout = req.getLast();
-        var resp = redisListCore.blpop(key, Integer.parseInt(timeout));
+        var resp = redisListCore.blpop(key, Double.parseDouble(timeout));
         if (resp == null) {
             RedisWriteProcessor.sendNull(outputStream);
         } else {
